🔨 로직 리팩토링
