+++ apps/web/src/components/RecipeCard/index.ts
@@ -1,0 +1,1 @@
+export { RecipeCard } from './RecipeCard';
